Move testdata up one level, no point in being data/testdata
