+++ exporter/exporterhelper/resource_to_label_test.go
@@ -18,7 +18,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	"go.opentelemetry.io/collector/internal/data/testdata"
+	"go.opentelemetry.io/collector/internal/testdata"
 )
 
 func TestConvertResourceToLabels(t *testing.T) {
