+++ exporter/loggingexporter/logging_exporter_test.go
@@ -23,7 +23,7 @@ import (
 
 	"go.opentelemetry.io/collector/config/configmodels"
 	"go.opentelemetry.io/collector/consumer/pdata"
-	"go.opentelemetry.io/collector/internal/data/testdata"
+	"go.opentelemetry.io/collector/internal/testdata"
 )
 
 func TestLoggingTraceExporterNoErrors(t *testing.T) {
