+++ processor/fanoutconnector_test.go
@@ -23,7 +23,7 @@ import (
 
 	"go.opentelemetry.io/collector/consumer"
 	"go.opentelemetry.io/collector/consumer/consumertest"
-	"go.opentelemetry.io/collector/internal/data/testdata"
+	"go.opentelemetry.io/collector/internal/testdata"
 )
 
 func TestTracesProcessorNotMultiplexing(t *testing.T) {
