+++ translator/internaldata/oc_to_metrics_test.go
@@ -24,7 +24,7 @@ import (
 
 	"go.opentelemetry.io/collector/consumer/consumerdata"
 	"go.opentelemetry.io/collector/consumer/pdata"
-	"go.opentelemetry.io/collector/internal/data/testdata"
+	"go.opentelemetry.io/collector/internal/testdata"
 )
 
 func TestOCToMetrics(t *testing.T) {
