+++ translator/trace/zipkin/traces_to_zipkinv2_test.go
@@ -25,8 +25,8 @@ import (
 
 	"go.opentelemetry.io/collector/consumer/pdata"
 	otlptrace "go.opentelemetry.io/collector/internal/data/opentelemetry-proto-gen/trace/v1"
-	"go.opentelemetry.io/collector/internal/data/testdata"
 	"go.opentelemetry.io/collector/internal/goldendataset"
+	"go.opentelemetry.io/collector/internal/testdata"
 )
 
 func TestInternalTracesToZipkinSpans(t *testing.T) {
